requirements.txt
betterproto==1.2.5
black==22.10.0
click==8.1.3
Flask==2.2.2
flask-grpc==0.1.1
grpcio==1.50.0
grpcio-tools==1.50.0
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
protobuf==4.21.8
PyYAML==6.0
six==1.16.0
stringcase==1.2.0
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
